// apps/end-user-app/src/types/finance.ts
export interface Asset {
  reference: string
  name: string
  symbol: string
  decimals: number
  is_settleable: boolean
}
// ...
export interface Instrument {
  reference: string
  name: string
  quantity_decimals: number
  px_decimals: number
  instrument_type: string
  base_asset_reference?: string
  quote_asset_reference?: string
  settlement_asset_reference?: string
  underlying_asset_reference?: string
  staking_asset_reference?: string
  yielding_asset_reference?: string
}
// ...
export interface Portfolio {
  reference: string
  name: string
  description?: string
}

export interface CreatePortfolioFormInputs {
  name: string
  description?: string
}

export interface UpdatePortfolioFormInputs {
  name?: string
  description?: string
}

// LedgerEntry

export interface LedgerEntry {
  reference: string
  source_type: string
  entry_time: string
  entry_type: string
  settlement_amount_change: number
  settlement_asset_reference: string
  instrument_reference: string | null
  quantity_change: number | null
  fill_px: number | null
  remark: string | null
  parent_ledger_entry_reference: string | null
  children_ledger_entries: LedgerEntry[]
}

export interface CreateLedgerEntryFormInputs {
  entry_time: string
  entry_type: string
  settlement_amount_change: number
  settlement_asset_reference: string
  quantity_change: number | null
  instrument_reference: string | null
  fill_px: number | null
  remark: string | null
  parent_ledger_entry_reference: string | null
}

export interface UpdateLedgerEntryFormInputs {
  entry_time: string
  entry_type: string
  settlement_amount_change: number
  settlement_asset_reference: string
  quantity_change: number | null
  instrument_reference: string | null
  fill_px: number | null
  remark: string | null
  parent_ledger_entry_reference: string | null
}